requirements.txt
streamlit>=1.31.0
google-generativeai>=0.5.0
openai
groq>=0.9.0